review fixes

+++ LykkeApi2/Models/Deposits/SwiftDepositEmailModel.cs
@@ -2,6 +2,6 @@
 {
     public class SwiftDepositEmailModel
     {
-        public double Amount { set; get; }
+        public decimal Amount { set; get; }
     }
 }
